feat:wrote a small utility node function to generate a ts file that exports a map from string to all the game Script classes present in the Scripts folder.
This generated ts file will be named scriptMap and be placed in src/Engine. (added this to gitignore as not a good idea to commit files that can be generated)

The node function must also be compiled using tsc in the Gen working directory before execution. (added this to gitignore as not a good idea to commit files that can be generated)