feat:script system overhaul
Added a new JSON module in common.d.ts, json is the format in which a game scene will be described.

Removed a lot of junk code from main. also added in an actual game loop.

Minor refactoring in actor, will have 0 code impact.

CScriptComponent is completely overhauled, onTick and onStart are no longer to be in composition with it. instead it is expected that the CScriptComponent class will be specialized and these two methods be overridden. The analysis was that this will help me to provide the scripting api easily (protected inside base). Bad overrides like an override for getComponentType is dangerous but individual scripts should not be that big and as such not unwieldy if anything goes wrong.

Not much has changed in the ScriptProcessor. onTick will now be called on the next frame after onStart.

Wrote a small script testActorScript to test scripting in the engine.