feat: resource mapping for now is manual.
Added comments and new entries into the gitignore, removed scriptMap as it will no longer be generated automatically.

Added new word 'npcs' into the settings.json

Manually wrote the scriptMap files, these files need to be updated with new resources as and when needed, the scene loader cannot access the resource folder, it will only look at the maps

Added couple of new interfaces to map the shape of a scene they are incredibly basic and will be worked on in the future.