feat:life of an actor
Cleared up main.

Processors no longer own actors and as such don't require components to register themselves with processors.

The interface to change scenes from a script is now defined in the Scene.ts file (the method will be implemented in the future).

GlobalComponentStore's api is now improved, can now efficiently get components of a certain type. also added comments for readability.

Expanded the responsibility of GlobalState, it is now also responsible for performing an operation on all processors at once.

Processor no longer owns an actor, only the GlobalComponentStore does, to that end removed the register and unregister api, instead a new getActorsInConsideration api is provided which children can override and call to see which actors to process.

Script and Sprite processors are refactored to take into account changes in Processor