feat:able to load in tile layers from a PackedScene
Upgraded the project to ES2022, this mainly done for better API on types like arrays.

Added new words mostly used by TILED into the settings.json.

Main.ts now loads the default scene on startup and also instantiates a camera actor which the user can use to pan around the screen.

Also moved process calls from GlobalState to main, this is done to avoid circular dependencies.

Fix made in CTransformComponent, the getScale and setScale API were wrongly implemented.

Fix made to resource names in SceneMap

Added the CameraActorScript to scriptMap.

Started implementing the loadScene API, now able to load TileLayers from PackedScene. changed the loadScene API to now take in string which will be looked up in the resource map for the IPackedScene object.

Added more API to the interfaces IPackedScene, IPackedSceneLayer, IPackedTileset.

Also added in a new enum EnumPackedSceneLayerType, this is useful to determine wether a scene layer is made up of discrete objects or a cluster of tiles represented by their gids in an array.

Added a new interface ISceneObject to map out what a discrete object in an object layer might look like.

GlobalComponent store could not use a weak map with string. will make changes to the global component store as part of a future commit.

Moved out processor handling from GlobalState due to circular dependency problems.

loadScene Api now takes in a string so no point storing IPackedScene in GlobalState, made changes to that end.

SpriteProcessor now does not compulsorily require a main camera to work. if a main camera does not exist, it will create and use an identity matrix as its main camera.

SpriteProcessor is also now caching projectionMatrix, this is not an issue as currently there is no api to change said projection while the engine is running.

SpriteProcessor's GPUFragmentState was changed by adding a blending function to allow drawing of transparent objects on screen.